Remove size_t's and fix warnings in layer/

+++ layer/FlexiNoteLayer.h
@@ -46,7 +46,7 @@ public:
     virtual QString getFeatureDescription(View *v, QPoint &) const;
 
     virtual bool snapToFeatureFrame(View *v, int &frame,
-                    size_t &resolution,
+                    int &resolution,
                     SnapType snap) const;
 
     virtual void drawStart(View *v, QMouseEvent *);
@@ -70,7 +70,7 @@ public:
 
     virtual bool editOpen(View *v, QMouseEvent *);
 
-    virtual void moveSelection(Selection s, size_t newStartFrame);
+    virtual void moveSelection(Selection s, int newStartFrame);
     virtual void resizeSelection(Selection s, Selection newSize);
     virtual void deleteSelection(Selection s);
